Extend data loading and mesh creation utilities for uniform AMR datasets
- Add `get_tree_size` function to calculate header and offset sizes for AMRVAC data files
- Implement `load_from_uarrays` function in multiple modules to create AMR datasets from uniform grid data
- Refactor header template and data loading utilities
- Update type conversions and error handling in mesh creation functions

+++ src/simesh/meshes/amr_mesh.py
@@ -857,8 +857,8 @@ def amrmesh_from_uniform(nw_arrays:np.ndarray, w_names, xmin, xmax, block_nx):
 
     assert isinstance(nw_arrays, np.ndarray), "nw_arrays must be a numpy array"
     assert len(nw_arrays.shape) == 4, "nw_arrays must be a 4D array"
-    domain_nx = np.array(nw_arrays.shape[:3])
-    block_nx = np.array(block_nx)
+    domain_nx = np.array(nw_arrays.shape[:3]).astype(int)
+    block_nx = np.array(block_nx).astype(int)
 
     assert nw_arrays.shape[3] == len(w_names), "nw_arrays and w_names must have the same length"
